# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_Astervia_wacraft-core_src_common_model.DescriptiveError:
    properties:
      content:
        description: Additional information about the error provided by the service
          that generated it. E. g. if an ORM returns an error, it will be propagated
          here as the raw error.
      message:
        description: Short text describing the error as best as we can for humans.
        type: string
    type: object
...
info:
  contact:
    email: wacraft@astervia.tech
    name: Astervia Dev Team
    url: https://github.com/Astervia
  description: Backend server for the wacraft project. Handles WhatsApp Cloud API
    operations, including message sending, receiving, and webhook handling.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: wacraft Server API
  version: 0.1.0
paths:
...
  /campaign/message/count:
    get:
      consumes:
      - application/json
      description: Returns the total number of campaign messages that match the given
        filters.
      parameters:
      - in: query
        name: campaign_id
        type: string
      - default: desc
        enum:
        - asc
        - desc
        in: query
        name: created_at
        type: string
        x-enum-varnames:
        - Asc
        - Desc
      - in: query
        name: created_at_geq
        type: string
      - in: query
        name: created_at_leq
        type: string
      - description: The unique identifier
        in: query
        name: id
        type: string
      - in: query
        name: message_id
        type: string
      - default: desc
        enum:
        - asc
        - desc
        in: query
        name: updated_at
        type: string
        x-enum-varnames:
        - Asc
        - Desc
      - in: query
        name: updated_at_geq
        type: string
      - in: query
        name: updated_at_leq
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Number of campaign messages
          schema:
            type: integer
        "400":
          description: Invalid query
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
      security:
      - ApiKeyAuth: []
      summary: Count campaign messages
      tags:
      - Campaign Message
  /campaign/message/count/sent:
    get:
      consumes:
      - application/json
      description: Returns the number of campaign messages where the message ID is
        not null (sent messages).
      parameters:
      - in: query
        name: campaign_id
        type: string
      - default: desc
        enum:
        - asc
        - desc
        in: query
        name: created_at
        type: string
        x-enum-varnames:
        - Asc
        - Desc
      - in: query
        name: created_at_geq
        type: string
      - in: query
        name: created_at_leq
        type: string
      - description: The unique identifier
        in: query
        name: id
        type: string
      - in: query
        name: message_id
        type: string
      - default: desc
        enum:
        - asc
        - desc
        in: query
        name: updated_at
        type: string
        x-enum-varnames:
        - Asc
        - Desc
      - in: query
        name: updated_at_geq
        type: string
      - in: query
        name: updated_at_leq
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Number of sent campaign messages
          schema:
            type: integer
        "400":
          description: Invalid query
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
      security:
      - ApiKeyAuth: []
      summary: Count sent campaign messages
      tags:
      - Campaign Message
  /campaign/message/count/unsent:
    get:
      consumes:
      - application/json
      description: Returns the number of campaign messages where the message ID is
        null (unsent messages).
      parameters:
      - in: query
        name: campaign_id
        type: string
      - default: desc
        enum:
        - asc
        - desc
        in: query
        name: created_at
        type: string
        x-enum-varnames:
        - Asc
        - Desc
      - in: query
        name: created_at_geq
        type: string
      - in: query
        name: created_at_leq
        type: string
      - description: The unique identifier
        in: query
        name: id
        type: string
      - in: query
        name: message_id
        type: string
      - default: desc
        enum:
        - asc
        - desc
        in: query
        name: updated_at
        type: string
        x-enum-varnames:
        - Asc
        - Desc
      - in: query
        name: updated_at_geq
        type: string
      - in: query
        name: updated_at_leq
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Number of unsent campaign messages
          schema:
            type: integer
        "400":
          description: Invalid query
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Astervia_wacraft-core_src_common_model.DescriptiveError'
      security:
      - ApiKeyAuth: []
      summary: Count unsent campaign messages
      tags:
      - Campaign Message
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"